updated samples for currenty supported generator

+++ samples/client/petstore/kotlin/src/main/kotlin/io/swagger/client/apis/StoreApi.kt
@@ -86,7 +86,7 @@ class StoreApi(basePath: kotlin.String = "http://petstore.swagger.io/v2") : ApiC
     * @return Order
     */
     @Suppress("UNCHECKED_CAST")
-    fun getOrderById(orderId: kotlin.Long) : Order {
+    fun getOrderById(orderId: kotlin.Int) : Order {
         val localVariableBody: kotlin.Any? = null
         val localVariableQuery: MultiValueMap = mapOf()
         val localVariableHeaders: kotlin.collections.Map<kotlin.String,kotlin.String> = mapOf()
